# Improve performance of buzhash