@Kubuxu 2019-11-06
I've confirmed with go-ipfs that output is the same.